Make sure the default character set for the database is UTF-8
The default character set for MySQL database is still latin1 with a
default collation of latin1_swedish_ci.  When creating the new database
specifying a character set of 'utf8' should ensure that the character
set of the tables and columns created is also UTF-8.  (With the
character set of a column as latin1, you can't store Farsi text, for
instance.)