Move application_name definition and add a new 'debug' setting
The motivation of this change is to allow us to deploy both a staging
site and a production site to the same server if we want.
application_name might be 'writeit-live' for the production site, for
example.  The 'debug' variable controls whether the Django project
settings have DEBUG true or not.

You now would use a different playbook to deploy to live or staging; for
deploying to staging, you'd use:

   site.yml / vagrant.yml

... while to deploy to live, you'd use:

   site-live.yml / vagrant.yml

This commit also renames the host group in the inventory to something
more clear; we may be deploying both a production site and a staging
site on the EC2 server created for this project, so to put it in a group
called "staging" would be confusing.